requirements.txt
Fabric==1.8.2
PyYAML==3.10
ecdsa==0.10
paramiko==1.12.2
pycrypto==2.6.1
wsgiref==0.1.2
